server/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
anyio==3.5.0
asn1crypto==1.4.0
async-timeout==4.0.2
attrs==21.4.0
Authlib==0.15.5
base58==2.1.1
bitarray==1.2.2
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
coincurve==17.0.0
cryptography==36.0.1
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.0.2
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
hexbytes==0.2.2
httpcore==0.14.7
httpx==0.22.0
idna==3.3
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.0
protobuf==3.19.4
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pyrsistent==0.18.1
pysha3==1.0.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
rlp==2.0.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
tomli==2.0.1
toolz==0.11.2
typing_extensions==4.0.1
urllib3==1.26.8
varint==1.0.2
web3==5.28.0
websockets==9.1
Werkzeug==2.0.3
WTForms==3.0.1
yarl==1.7.2
